requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile setup.py
#
absl-py==0.9.0            # via tensorboard, tensorflow
astunparse==1.6.3         # via tensorflow
beautifulsoup4==4.9.1     # via wikipedia
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via matplotlib, requests
cffi==1.14.2              # via cld2-cffi
chardet==3.0.4            # via requests
cld2-cffi==0.1.4          # via lidtk (setup.py)
click==7.1.2              # via lidtk (setup.py), nltk
cycler==0.10.0            # via matplotlib
detectlanguage==1.4.0     # via lidtk (setup.py)
fuzzywuzzy==0.18.0        # via lidtk (setup.py)
gast==0.3.3               # via tensorflow
google-auth-oauthlib==0.4.1  # via tensorboard
google-auth==1.20.1       # via google-auth-oauthlib, tensorboard
google-pasta==0.2.0       # via tensorflow
grpcio==1.31.0            # via tensorboard, tensorflow
h5py==2.10.0              # via keras, keras-applications, lidtk (setup.py), tensorflow
idna==2.10                # via requests
joblib==0.16.0            # via nltk, scikit-learn
keras-applications==1.0.8  # via keras
keras-preprocessing==1.1.2  # via keras, tensorflow
keras==2.3.1              # via lidtk (setup.py)
kiwisolver==1.2.0         # via matplotlib
langdetect==1.0.8         # via lidtk (setup.py)
langid==1.1.6             # via lidtk (setup.py)
markdown==3.2.2           # via tensorboard
matplotlib==3.3.1         # via lidtk (setup.py), seaborn
nltk==3.5                 # via lidtk (setup.py)
numpy==1.18.5             # via h5py, keras, keras-applications, keras-preprocessing, langid, lidtk (setup.py), matplotlib, opt-einsum, pandas, scikit-learn, scipy, seaborn, tensorboard, tensorflow
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.3.0         # via tensorflow
pandas==1.1.0             # via seaborn
pillow==7.2.0             # via matplotlib
progressbar2==3.51.4      # via lidtk (setup.py)
protobuf==3.13.0          # via tensorboard, tensorflow
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib, pandas
python-levenshtein==0.12.0  # via lidtk (setup.py)
python-utils==2.4.0       # via progressbar2
pytz==2020.1              # via pandas
pyyaml==5.3.1             # via keras, lidtk (setup.py)
regex==2020.7.14          # via nltk
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.24.0          # via detectlanguage, requests-oauthlib, tensorboard, wikipedia
rsa==4.6                  # via google-auth
scikit-learn==0.23.2      # via lidtk (setup.py)
scipy==1.4.1              # via keras, lidtk (setup.py), scikit-learn, seaborn, tensorflow
seaborn==0.10.1           # via lidtk (setup.py)
six==1.15.0               # via absl-py, astunparse, cld2-cffi, cycler, google-auth, google-pasta, grpcio, h5py, keras, keras-preprocessing, langdetect, progressbar2, protobuf, python-dateutil, python-utils, tensorboard, tensorflow
soupsieve==2.0.1          # via beautifulsoup4
tensorboard-plugin-wit==1.7.0  # via tensorboard
tensorboard==2.3.0        # via tensorflow
tensorflow-estimator==2.3.0  # via tensorflow
tensorflow==2.3.0         # via lidtk (setup.py)
termcolor==1.1.0          # via tensorflow
threadpoolctl==2.1.0      # via scikit-learn
tqdm==4.48.2              # via nltk
urllib3==1.25.10          # via requests
werkzeug==1.0.1           # via tensorboard
wheel==0.35.1             # via astunparse, tensorboard, tensorflow
wikipedia==1.4.0          # via lidtk (setup.py)
wrapt==1.12.1             # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
